+++ lib/presentation/constant/size/constant_size.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
 const SizedBox kHeight = SizedBox(height: 30);
+const SizedBox kHeight50 = SizedBox(height: 50);
 const SizedBox kHeight20 = SizedBox(height: 20);
 const SizedBox kHeight10 = SizedBox(height: 10);
