+++ lib/presentation/otp_screen/otp_screen.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+
+import '../constant/color/colors.dart';
+import '../constant/font/google_font.dart';
+import '../constant/size/constant_size.dart';
+
+class OtpScreen extends StatelessWidget {
+  const OtpScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: appBgColor,
+      appBar: AppBar(
+        backgroundColor: appBgColor,
+        elevation: 0,
+        iconTheme: const IconThemeData(
+          color: appBlackColor,
+        ),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 25),
+        child: Column(
+          children: <Widget>[
+            Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: <Widget>[
+                kHeight,
+                Text(
+                  'Enter OTP.',
+                  style: GoogleFont.headTextStyle,
+                ),
+                kHeight20,
+                Center(child: Image.asset('assets/images/otp.png')),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
