+++ lib/presentation/signup_screen/signup_screen.dart
@@ -1,0 +1,100 @@
+import 'package:flutter/material.dart';
+import '../constant/color/colors.dart';
+import '../constant/font/google_font.dart';
+import '../constant/size/constant_size.dart';
+import '../otp_screen/otp_screen.dart';
+import '../widget/eleveted_button_widget/elevated_button_widget.dart';
+import '../widget/text_field_widget/text_field_widget.dart';
+
+class SignupScreen extends StatelessWidget {
+  SignupScreen({super.key});
+  final GlobalKey _formKey = GlobalKey<FormState>();
+  final TextEditingController _sampleController = TextEditingController();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: appBgColor,
+      appBar: AppBar(
+        backgroundColor: appBgColor,
+        elevation: 0,
+        iconTheme: const IconThemeData(
+          color: appBlackColor,
+        ),
+      ),
+      body: SingleChildScrollView(
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 25),
+          child: Column(
+            children: <Widget>[
+              Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: <Widget>[
+                  kHeight,
+                  Text(
+                    'Create Account.',
+                    style: GoogleFont.headTextStyle,
+                  ),
+                  Text(
+                    'Signup with your phone number',
+                    style: GoogleFont.subHeadTextStyle,
+                  ),
+                  kHeight20,
+                  Center(child: Image.asset('assets/images/create.png')),
+                ],
+              ),
+              kHeight50,
+              Form(
+                key: _formKey,
+                child: Column(
+                  children: <Widget>[
+                    TextFieldWidget(
+                      controller: _sampleController,
+                      hintText: 'Username',
+                    ),
+                    TextFieldWidget(
+                      controller: _sampleController,
+                      hintText: 'Phone',
+                    ),
+                    TextFieldWidget(
+                      controller: _sampleController,
+                      hintText: 'Password',
+                    ),
+                    kHeight10,
+                    ElvatedButtonWidget(
+                      buttonText: 'Get OTP',
+                      onpressed: () {
+                        Navigator.of(context).push(
+                          MaterialPageRoute<OtpScreen>(
+                            builder: (BuildContext _) => OtpScreen(),
+                          ),
+                        );
+                      },
+                    ),
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        Text(
+                          'Already have an account?',
+                          style: GoogleFont.subHeadTextStyle,
+                        ),
+                        TextButton(
+                          onPressed: () {},
+                          child: Text(
+                            'Sign In',
+                            style: GoogleFont.textButtonStyle,
+                          ),
+                        )
+                      ],
+                    ),
+                    kHeight,
+                  ],
+                ),
+              )
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
